move regrid mask tests to their own file
